backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2
openai==1.3.0
redis==5.0.1
python-dotenv==1.0.0
python-multipart==0.0.6
asyncio-mqtt==0.13.0
aiofiles==23.2.1
websockets==12.0
